pyproject.toml
[project]
name = "IBL-fiberphotometry-to-nwb"
version = "0.0.1"
description = "NWB conversion scripts, functions, and classes for IBL-fiberphotometry  conversion"
readme = "README.md"
authors = [{ name = "CatalystNeuro", email = "ben.dichter@catalystneuro.com" }]
maintainers = [{ name = "CatalystNeuro", email = "ben.dichter@catalystneuro.com" }]
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]

dependencies = [
  "neuroconv",
  "nwbinspector",
  "ONE-api",
  "sparse",
  "ndx-ibl-bwm@git+https://github.com/int-brain-lab/ndx-ibl-bwm.git",
  "ibl-to-nwb@git+https://github.com/h-mayorquin/IBL-to-nwb.git@heberto_conversion",
  "ndx-ibl@git+https://github.com/catalystneuro/ndx-ibl.git",
  "ndx-fiber-photometry",
]

[project.urls]
Repository="https://github.com/catalystneuro/IBL-fiberphotometry-to-nwb"

[project.optional-dependencies]
fiber_photometry = [
  "neuroconv==0.7.0"   # Pinned dependencies to specific versions
]

[dependency-groups]
dev = [
  "pre-commit",
  "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "*.yaml",
  "*.yml",
  "*.json",
]  # This includes the yaml (usually metadata) and any schemas you might need on both the sdist and wheel

[tool.hatch.build.targets.wheel]
packages = ["src/ibl_fiberphotometry_to_nwb"]

[tool.hatch.build.targets.sdist]
packages = ["src/ibl_fiberphotometry_to_nwb"]

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.toml
  |\.yml
  |\.md
  |\.txt
  |\.sh
  |\.git
  |\.ini
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.codespell]
skip = '.git*,*.pdf,*.css'
check-hidden = true
ignore-words-list = 'assertin'

[tool.ruff]

[tool.ruff.lint]
select = [
    "F401",  # Unused import
    "I",  # All isort rules
    "UP006",  # non-pep585 annotation (tuple, list -> tuple, list)
    "UP007",  # non-pep604 annotation (Union[x, y] -> X | Y)
    "UP045",  # non pep604 annotation (Optional[x] -> x | None)
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
relative-imports-order = "closest-to-furthest"
known-first-party = ["ibl_fiberphotometry_to_nwb"]